gjøre knappen gjenbrukbar og sende videre til steg
